// packages/sdk/src/interfaces/types.ts
export enum MessageDirection {
  L1_TO_L2,
  L2_TO_L1,
}
// ...
export interface NFTBridgeMessage {
  direction: MessageDirection
  from: string
  to: string
  l1Token: string
  l2Token: string
  extraData: string
  tokenId: BigNumber
  logIndex: number
  blockNumber: number
  transactionHash: string
}
// ...
export type AddressLike = string | Contract

/**
 * Stuff that can be coerced into a number.
 */
export type NumberLike = string | number | BigNumber